refactor:  add estimatedAnnualRewards to staking input hook
It adds the estimatedAnnualRewards state and the calculateEstimatedAnnualRewards function to calculate and display the estimated annual rewards based on the staking amount and the annual reward rate.
